이미지 처리 기능 개선 및 코드 구조 변경
`ImagePipeline.fs`에서 사용되지 않는 메서드 제거,
`BoxedZoneEditAdaptor.fs`에 새로운 타입 추가 및 메서드 수정,
프로젝트 파일에서 컴파일 포함 변경,
`LuticaSKIDAPI.fs`에서 이미지 처리 명령어 확장,
`TextureImageProcessing.fs`에서 부분 이미지 처리 기능 추가,
`SKIDToolFunction.fs`에 이미지 크롭 함수 추가,
`StructTypes.fs`에서 타입 생성자 수정,
새로운 `SKIDAPICommands.fs` 파일 추가,
`SKIDToolFunctionTests.fs`에 테스트 추가로 코드 안정성 향상.

+++ LuticaSKID/Schema/StructTypes.fs
@@ -93,8 +93,8 @@ module public StructTypes =
             SKIDPixelVector2(a.x * b, a.y * b)
         static member (*) (a: SKIDPixelVector2, b: SKIDPixelVector2) =
             SKIDPixelVector2(a.x * b.x, a.y * b.y)
-        static member (/) (a: SKIDPixelVector2, b: float32): Result<SKIDPixelVector2, DivideByZeroException> =
-            if b = 0.0f then Error (DivideByZeroException "Division by zero") else Ok (SKIDPixelVector2(a.x / b, a.y / b))
+        static member (/) (a: SKIDPixelVector2, b: int): Result<SKIDPixelVector2, DivideByZeroException> =
+            if b = 0 then Error (DivideByZeroException "Division by zero") else Ok (SKIDPixelVector2(a.x / b, a.y / b))
 
     [<Class>]
     type ColorSpaceBoundary =
@@ -169,11 +169,12 @@ module public StructTypes =
         new(pixels: SKIDColor[], width: int, height: int) =
             { pixels = pixels; width = width; height = height }
 
+
     [<Class>]
     type ImageProcessInput<'TSetting> = 
         val image: SKIDImage
         val config: 'TSetting option
-        new(image:SKIDImage, config: 'TSetting option) = { image = image; config = config}
+        new (image:SKIDImage, config: 'TSetting option) = { image = image; config = config }
         new (image:SKIDImage) = { image = image; config = None}
     
     type ImageProcessingOrder =
